requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
requests==2.31.0
openai==1.23.2
httpx==0.27.0
